perf(c): add C version of select waiting function
'Wait' and 'Ready' enums moved into the '_enums' module to avoid a
circular import.
